+++ main.go
@@ -1,7 +1,9 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"os"
 	"time"
 
 	"github.com/sirupsen/logrus"
@@ -11,17 +13,31 @@ func main() {
 	logrus.Infof("Starting up")
 
 	hourOfDay := time.Now().Hour()
-	greeting := getGreeting(hourOfDay)
+	greeting, err := getGreeting(hourOfDay)
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+
 	fmt.Println(greeting)
 	logrus.Infof("Normal exit")
 }
 
-func getGreeting(hour int) string {
-	if hour < 12 {
-		return "Good morning"
-	} else if hour < 18 {
-		return "Good afternoon"
+func getGreeting(hour int) (string, error) {
+	var message string
+
+	if hour < 7 {
+		err := errors.New("Too early for Gophers")
+		return message, err
 	}
 
-	return "Good evening"
+	if hour < 12 {
+		message = "Good morning"
+	} else if hour < 18 {
+		message = "Good afternoon"
+	} else {
+		message = "Good evening"
+	}
+
+	return message, nil
 }